added test on config existence

+++ tests/tests.sh
@@ -2,10 +2,11 @@
 
 cd ./apis_services/diskspacechecks
 echo HOME is $HOME
-ls -ltra $HOME
-mkdir $HOME/.ssh
-cp /sshkeys/* $HOME/.ssh
-chmod 0700 $HOME/.ssh && chmod 0600 $HOME/.ssh/*
+if [ ! -f $HOME/.ssh/config ] ; then 
+    mkdir $HOME/.ssh
+    cp /sshkeys/* $HOME/.ssh
+    chmod 0700 $HOME/.ssh && chmod 0600 $HOME/.ssh/*
+fi 
 pip install -r requirements.txt
 pip install pytest pytest-cov
 pytest -v .
